Change simulation trajectory
Since we must be able to change the relative pose between IMU and
camera, the trajectory stored in the dataset must be the one used by
the camera. Otherwise there will be no or few visible landmarks since
these are tied to the trajectory.

Instead, we create a new simulation trajectory such that the applying the
relative pose on the new trajectory yields the old one.